@@ -35,7 +35,7 @@ static size_t hash(Eigen::Vector2d state) {
  * You'll probably want to override the transitionValidator callback
  * if your 2d plane has any obstacles.
  */
-RRT::Tree<Eigen::Vector2d>* TreeFor2dPlane(
+std::shared_ptr<RRT::Tree<Eigen::Vector2d>> TreeFor2dPlane(
     std::shared_ptr<StateSpace<Eigen::Vector2d>> stateSpace,
     Eigen::Vector2d goal, double step);
 
